fix: cpu limit can be a decimal number

+++ src/main/java/eu/divum/divumbackend/dtos/minecraftserverinstance/MinecraftServerInstanceConfiguration.java
@@ -8,7 +8,7 @@ import java.util.List;
 public class MinecraftServerInstanceConfiguration {
     private int memoryLimit;
 
-    private int cpuCoresLimit;
+    private float cpuCoresLimit;
 
     private boolean eula;
